--- recon/dogpile.py
@@ -1,34 +1,0 @@
-#!/usr/bin/env python
-# -*- coding:utf-8 -*- 
-#
-# @name   : Infoga - Email Information Gathering
-# @url    : http://github.com/m4ll0k
-# @author : Momo Outaadi (m4ll0k)
-
-from lib.output import *
-from lib.request import *
-from lib.parser import *
-
-class Dogpile(Request):
-	def __init__(self,target):
-		Request.__init__(self)
-		self.target = target
-
-	def search(self):
-		test('Searching "%s" in DogPile...'%(self.target))
-		url = "http://www.dogpile.com/search/web?qsi=0&q=%40{target}".format(
-			target=self.target)
-		try:
-			resp = self.send(
-				method = 'GET',
-				url = url,
-				headers = {
-							'Host':'www.dogpile.com'
-				}
-				)
-			return self.getemail(resp.content,self.target)
-		except Exception as e:
-			pass
-
-	def getemail(self,content,target):
-		return parser(content,target).email()
